Ronda de fixes post-testing sobre las capacidades de items pasivos:
force door con formula NxM + UI de breakdown, gate FamilyExclusive
(fortuna excluyente), Instinto = pasiva del Warrior como starting item,
Moneda Maldita pondera malditos en el altar, canal float del base damage
override (Furia 0.25/ronda) y fortuna pagando 5/roll con feedback.

Conflictos resueltos a mano en las 3 tablas de localizacion (Content
Shared Data / Content_es / Content_en): ambos lados apendearon entries
al final — develop sumo keys de bosses/curses/intents y la rama las de
corazon.fortuna. Se conservaron ambos bloques.

+++ Assets/Scripts/Rollgeon/Combat/Damage/DamageBreakdown.cs
@@ -16,15 +16,18 @@ namespace Rollgeon.Combat.Damage
         // ── Términos de N (aditivos) ──────────────────────────────────────
         /// <summary>Base del combo (0 en el fallback sin combo con detalle de bag).</summary>
         public int ComboBase;
-        /// <summary>dmg_base_PJ — <c>Attack.Value</c>.</summary>
-        public int AttackBase;
+        /// <summary>dmg_base_PJ — <c>Attack.Value</c>, o el base damage override (float:
+        /// Furia Contenida acumula 0.25/ronda y la fracción viaja hasta el redondeo final).</summary>
+        public float AttackBase;
         /// <summary>bonos_PJ — <c>Attack.ModifiedValue − Attack.Value</c>.</summary>
         public int AttackBonus;
         /// <summary>Σ caras de los dados contribuyentes.</summary>
         public int FacesSum;
         /// <summary>Σ BonusComboDamage de los 3 canales de scratch.</summary>
         public int AdditiveBonus;
-        public int N;
+        /// <summary>Suma de los términos aditivos — float: solo <see cref="AttackBase"/>
+        /// puede aportar fracción; el redondeo único vive en <c>RoundNxM(N, M)</c>.</summary>
+        public float N;
 
         // ── Términos de M (multiplicativos) ───────────────────────────────
         /// <summary>Producto de ComboDamageMultiplier de los 3 canales de scratch.</summary>
